Move the visitor inside repsonse because it's shared by a couple plugins
