+++ apollo-router/src/plugins/demand_control/cost_calculator/mod.rs
@@ -1,5 +1,4 @@
 mod directives;
-pub(crate) mod schema_aware_response;
 pub(crate) mod static_cost;
 
 use crate::plugins::demand_control::DemandControlError;
